GET PICTURES WORKS WITH API!
