+++ omhaw16/mvc/app/controllers/apiController.php
@@ -23,6 +23,31 @@ public function users() {
             }
         }
 
+public function pictures($userID) {
+
+			if ($this->get()) {
+				$objectOfPics = $this->model('Pictures');
+				$picJSON = $objectOfPics->getMyPosts($userID);
+
+//				echo json_encode($picJSON);
+
+				echo json_encode($picJSON);
+//				echo json_encode($objectOfPics);
+
+			}
+
+			else if ($this->post($userID,$username,$password)) {
+				$objectOfPics = $this->model('Pictures');
+				$objectOfUser = $this->model('User');
+
+				if ($objectOfUser->login($username,$password)) { 
+				$objectOfPics->uploadPic($userID,$imgname,$imgtitle,$imgdesc);
+			} else {
+				echo "Couldn't log in with " . $username . "and " . $password;
+			}
+		}
+}
+
 }
 
 ?>
